Relocate the current user service.
